+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All changes to XM WEB CLIENT project will be documented in this file.
 
 ## Version 4
+### [4.28] - 2022-05-24
+### Fix
+- Fix bug can not scroll to origin mess when view search/ label in all rooms when the room in sidebar is not rendered yet
+- Fix bug can not add tag person in begin of new line break
+- Fix bug tag broken when add emoji 
+- Don't show last seen time for channels
+
 ### [4.27] - 2022-05-20
 ### Update
 - Now "Send with Markdown" is set to true when markdown format is detected.